Missed an int64_t in the peek() conversion.

+++ toys/posix/file.c
@@ -34,7 +34,7 @@ static void do_elf_file(int fd)
 {
   unsigned endian = toybuf[5], bits = toybuf[4]-1, i, j, dynamic = 0,
            stripped = 1, phentsize, phnum, shsize, shnum, bail = 0, arch;
-  int64_t (*elf_int)(void *ptr, unsigned size) = (endian==2)?peek_be:peek_le;
+  long long (*elf_int)(void *ptr, unsigned size) = (endian==2)?peek_be:peek_le;
   char *map = MAP_FAILED;
   unsigned long phoff, shoff;
